fail when template is not given
Details:

The default template was an empty string. When not specified, the code
would crash without a meaningful error message. If this is a required
option, then we should check that it's being specified (done
now). Otherwise we should have a default template (doesn't make a lot
of sense to me).