modified template goes to the directory from which we get the template
Details:

without this change, the modified template is created in the current
directory; by default, pwd is not in pythonpath. The directory where
the script and the template are, however is in pytonpath.  This change
fixes the error

  No module named mytemplate

We might need to change it if we decide to have one dir for templates
and one for actual modules.