re-implement addBG -> getBackground
Details:
simplify the function (no need to loop over the json file line by
line). Specify the variation name as an argument, so that this
function can work also with the syst case. Clarify the fact that there
could be several backgrounds in the json file (rename 'BG' ->
'background1').

I still think that storing N backgrounds in the json is a bad idea.
Fbu does not care, and just wants one total background yield for now.